+++ shell/edit/networking.k8s.io.networkpolicy/__tests__/PolicyRuleTarget.spec.ts
@@ -2,46 +2,7 @@ import { mount } from '@vue/test-utils';
 import PolicyRuleTarget from '@shell/edit/networking.k8s.io.networkpolicy/PolicyRuleTarget';
 import { cleanHtmlDirective } from '@shell/plugins/clean-html-directive';
 import mock from '@shell/edit/networking.k8s.io.networkpolicy/__tests__/mock.json';
-
-class Selectors {
-  private wrapper;
-
-  constructor(wrapper) {
-    this.wrapper = wrapper;
-  }
-
-  get ipBlock() {
-    return this.wrapper.find('[data-testid=labeled-input-ip-block-selector]');
-  }
-
-  /**
-   * Namespace selector element; matches policies in OR condition, in conjunction with the other rules
-   */
-  get namespace() {
-    return this.wrapper.find('[data-testid=match-expression-namespace-selector]');
-  }
-
-  /**
-   * Pod selector element; matches policies in OR condition, in conjunction with the other rules
-   */
-  get pod() {
-    return this.wrapper.find('[data-testid=match-expression-pod-selector]');
-  }
-
-  /**
-   * Namespace and pod selector elements, matching policies in AND condition, within the same rule
-   */
-  get namespaceAndPod() {
-    return {
-      namespaceRule: this.wrapper.find('[data-testid=match-expression-namespace-and-pod-selector-ns-rule]'),
-      podRule:       this.wrapper.find('[data-testid=match-expression-namespace-and-pod-selector-pod-rule]'),
-    };
-  }
-
-  get ruleType() {
-    return this.wrapper.find('[data-testid=labeled-select-type-selector]');
-  }
-}
+import { PolicyRuleTargetSelectors } from '@shell/edit/networking.k8s.io.networkpolicy/__tests__/selectors';
 
 describe.each([
   'view',
@@ -80,7 +41,7 @@ describe.each([
 
       await wrapper.setProps({ value: { ipBlock } });
 
-      const selectors = new Selectors(wrapper);
+      const selectors = new PolicyRuleTargetSelectors(wrapper);
 
       // Check rule type selector
       expect(selectors.ruleType.vm.$data._value.value).toBe('ipBlock');
@@ -98,7 +59,7 @@ describe.each([
 
       await wrapper.setProps({ value: { namespaceSelector } });
 
-      const selectors = new Selectors(wrapper);
+      const selectors = new PolicyRuleTargetSelectors(wrapper);
 
       // Check rule type selector
       expect(selectors.ruleType.vm.$data._value.value).toBe('namespaceSelector');
@@ -123,7 +84,7 @@ describe.each([
 
       await wrapper.setProps({ value: { podSelector } });
 
-      const selectors = new Selectors(wrapper);
+      const selectors = new PolicyRuleTargetSelectors(wrapper);
 
       // Check rule type selector
       expect(selectors.ruleType.vm.$data._value.value).toBe('podSelector');
@@ -153,7 +114,7 @@ describe.each([
         }
       });
 
-      const selectors = new Selectors(wrapper);
+      const selectors = new PolicyRuleTargetSelectors(wrapper);
 
       // Check rule type selector
       expect(selectors.ruleType.vm.$data._value.value).toBe('namespaceAndPodSelector');
